Various Fixes
Fixed strtoul call so it is now just base 16. Everything else here is so it just makes it more consistent and easier to work with. added cases for peek/poke to work on locations above 9F, since hex digits were originally ignored. Added a break that was missing from a switch case. It does nothing but was preventing Windows compilation. Tested and ensured both Linux and Windows versions work.